+++ Nocturne/Features/Browse/AllAlbumsView.swift
@@ -41,7 +41,6 @@ struct AllAlbumsView: View {
                         }
                     }
                     .padding(Metrics.l)
-                    Color.clear.frame(height: 100)
                 }
             }
         }