+++ Nocturne/Features/Browse/PlaylistDetailView.swift
@@ -22,7 +22,6 @@ struct PlaylistDetailView: View {
                         playback.play(queue: songs, startAt: idx)
                     }
                 }
-                Color.clear.frame(height: 100)
             }
             .padding(.horizontal, Metrics.l)
         }